+++ src/admin/messages/de/cmsadmin.php
@@ -218,8 +218,8 @@ return [
     'module_permission_page' => 'Page Content',
 
 // 1.0.6
-	'page_update_actions_deepcopyastemplate_title' => 'Als Vorlage kopieren',
-	'page_update_actions_deepcopyastemplate_text' => 'Erstelle eine Vorlage der aktuellen Seite mit allen dazugehörigen Inhalten. Es werden dabei alle Inhalte in allen angelegten Sprachen von allen veröffentlichten Versionen kopiert.',
-	'page_update_actions_deepcopyastemplate_btn' => 'Vorlage erstellen',
-	'js_page_create_copy_as_template_success' => 'Die Vorlage wurde erfolgreich erstellt.',
+    'page_update_actions_deepcopyastemplate_title' => 'Als Vorlage kopieren',
+    'page_update_actions_deepcopyastemplate_text' => 'Erstelle eine Vorlage der aktuellen Seite mit allen dazugehörigen Inhalten. Es werden dabei alle Inhalte in allen angelegten Sprachen von allen veröffentlichten Versionen kopiert.',
+    'page_update_actions_deepcopyastemplate_btn' => 'Vorlage erstellen',
+    'js_page_create_copy_as_template_success' => 'Die Vorlage wurde erfolgreich erstellt.',
 ];