+++ src/admin/messages/pt/cmsadmin.php
@@ -219,8 +219,8 @@ return [
     'module_permission_page' => 'Conteúdo da Página',
 
 // 1.0.6
-	'page_update_actions_deepcopyastemplate_title' => 'Copiar como Modelo',
-	'page_update_actions_deepcopyastemplate_text' => 'Crie um modelo da página atual com todo o seu conteúdo. O modelo incluirá todos os idiomas, mas apenas a versão publicada.',
-	'page_update_actions_deepcopyastemplate_btn' => 'Criar modelo',
-	'js_page_create_copy_as_template_success' => 'O modelo foi criado.',
+    'page_update_actions_deepcopyastemplate_title' => 'Copiar como Modelo',
+    'page_update_actions_deepcopyastemplate_text' => 'Crie um modelo da página atual com todo o seu conteúdo. O modelo incluirá todos os idiomas, mas apenas a versão publicada.',
+    'page_update_actions_deepcopyastemplate_btn' => 'Criar modelo',
+    'js_page_create_copy_as_template_success' => 'O modelo foi criado.',
 ];