+++ src/admin/messages/ua/cmsadmin.php
@@ -220,8 +220,8 @@ return [
     'module_permission_page' => 'Page Content',
 
 // 1.0.6
-	'page_update_actions_deepcopyastemplate_title' => 'Copy as Template',
-	'page_update_actions_deepcopyastemplate_text' => 'Create a template of the current page with all its contents. They template will include all languages but only the published version.',
-	'page_update_actions_deepcopyastemplate_btn' => 'Create Template',
-	'js_page_create_copy_as_template_success' => 'The template has been created.',
+    'page_update_actions_deepcopyastemplate_title' => 'Copy as Template',
+    'page_update_actions_deepcopyastemplate_text' => 'Create a template of the current page with all its contents. They template will include all languages but only the published version.',
+    'page_update_actions_deepcopyastemplate_btn' => 'Create Template',
+    'js_page_create_copy_as_template_success' => 'The template has been created.',
 ];