+++ src/frontend/base/Controller.php
@@ -22,6 +22,72 @@ use luya\cms\frontend\Module;
 abstract class Controller extends \luya\web\Controller
 {
     /**
+     * @var string
+     * @since 1.0.8
+     */
+    const LINK_CANONICAL = 'linkCanonical';
+
+    /**
+     * @var string og:type key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_OG_TYPE = 'ogType';
+    
+    /**
+     * @var string twitter:card key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_TWITTER_CARD = 'twitterCard';
+    
+    /**
+     * @var string og:title key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_OG_TITLE = 'ogTitle';
+    
+    /**
+     * @var string twitter:title key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_TWITTER_TITLE = 'twitterTitle';
+    
+    /**
+     * @var string og:url key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_OG_URL = 'ogUrl';
+    
+    /**
+     * @var string twitter:url key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_TWITTER_URL = 'twitterUrl';
+
+    /**
+     * @var string description meta key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_DESCRIPTION = 'metaDescription';
+    
+    /**
+     * @var string og:description key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_OG_DESCRIPTION = 'ogDescription';
+    
+    /**
+     * @var string twitter:description key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_TWITTER_DESCRIPTION = 'twitterDescription';
+    
+    /**
+     * @var string keywords meta key which is used for meta registration. Use this constant in order to override the default implementation.
+     * @since 1.0.8
+     */
+    const META_KEYWORDS = 'metaKeywords';
+    
+    /**
      * Render the NavItem content and set several view specific data.
      *
      * @param integer $navItemId
@@ -102,17 +168,28 @@ abstract class Controller extends \luya\web\Controller
                 $this->view->title = $model->title_tag;
             }
         }
+
+
         
-        $this->view->registerMetaTag(['name' => 'og:title', 'content' => $this->view->title], 'fbTitle');
-        $this->view->registerMetaTag(['name' => 'og:type', 'content' => 'website'], 'ogType');
+
+        $this->view->registerMetaTag(['name' => 'og:type', 'content' => 'website'], self::META_OG_TYPE);
+        $this->view->registerMetaTag(['name' => 'twitter:card', 'content' => 'summary'], self::META_TWITTER_CARD);
+        
+        $this->view->registerMetaTag(['name' => 'og:title', 'content' => $this->view->title], self::META_OG_TITLE);
+        $this->view->registerMetaTag(['name' => 'twitter:title', 'content' => $this->view->title], self::META_TWITTER_TITLE);
+        
+        $this->view->registerLinkTag(['rel' => 'canonical', 'href' => Yii::$app->request->absoluteUrl], self::LINK_CANONICAL);
+        $this->view->registerMetaTag(['name' => 'og:url', 'content' => Yii::$app->request->absoluteUrl], self::META_OG_URL);
+        $this->view->registerMetaTag(['name' => 'twitter:url', 'content' => Yii::$app->request->absoluteUrl], self::META_TWITTER_URL);
 
         if (!empty($model->description)) {
-            $this->view->registerMetaTag(['name' => 'description', 'content' => $model->description], 'metaDescription');
-            $this->view->registerMetaTag(['name' => 'og:description', 'content' => $model->description], 'fbDescription');
+            $this->view->registerMetaTag(['name' => 'description', 'content' => $model->description], self::META_DESCRIPTION);
+            $this->view->registerMetaTag(['name' => 'og:description', 'content' => $model->description], self::META_OG_DESCRIPTION);
+            $this->view->registerMetaTag(['name' => 'twitter:description', 'content' => $model->description], self::META_TWITTER_DESCRIPTION);
         }
         
         if (!empty($model->keywords)) {
-            $this->view->registerMetaTag(['name' => 'keywords', 'content' => implode(", ", $currentMenu->keywords)], 'metaKeywords');
+            $this->view->registerMetaTag(['name' => 'keywords', 'content' => implode(", ", $currentMenu->keywords)], self::META_KEYWORDS);
         }
         
         if ($this->module->enableTagParsing) {